small tweaks to command structure & moved stuff around
